+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `Modified Date Control` will be documented in this file.
 ## v1.0.1
 
 - Fix issue with the plugin not persisting modified dates.
+- Fix issue with display of modified date.
 
 ## v1.0.0